Update script.js
added support for a "Stop at Top" function which will ease the circle's rotation back to its starting position with 0 degrees at the top, as well as cut off sound processing while the circle is easing back. Also made several changes to the orgnaization of the script for code maintainability.